+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 1. Update OKE Cluster functionality to be a container styles resource.
 2. Add Node Pool as a separate resource.
 3. Allow OKE Cluster to be dropped on compartment.
+4. Update Cache to include Node Pool Options.
 ### Bug Fixes